\name{makePackTable}
\alias{exportPackTable}
\alias{makePackTable}
\title{Create and export pack table}
\usage{
  makePackTable(sumtable)
}
\description{
  \code{makePackTable} creates a table with a list of
  spectra, their summary properties, and which pack they
  are in. \code{exportPackTable} then exports the table to
  CSV. The user can then select the appropriate spectra to
  compile by editing the table and adding OK markers to the
  spectra they want to select.
}
\author{
  Michele Stravs, Eawag
}
\seealso{
  readPackTable
}

